+++ docs/conf.py
@@ -19,9 +19,12 @@
 
 import sys
 
-# The name of the Pygments (syntax highlighting) style to use.
-from pygments.styles.default import DefaultStyle
-from pygments.token import Generic
+## The name of the Pygments (syntax highlighting) style to use.
+#from pygments.styles.default import DefaultStyle
+#from pygments.token import Generic
+
+sys.path.append(os.path.abspath("./_pygments"))
+pygments_style = 'style.HatchetStyle'
 
 import pkg_resources
 
@@ -80,12 +83,12 @@ master_doc = "index"
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 
-# modifications to the default style
-class HatchetStyle(DefaultStyle):
-    styles = DefaultStyle.styles.copy()
-    background_color = "#f4f4f8"
-    styles[Generic.Output] = "#355"
-    styles[Generic.Prompt] = "bold #346ec9"
+## modifications to the default style
+#class HatchetStyle(DefaultStyle):
+#    styles = DefaultStyle.styles.copy()
+#    background_color = "#f4f4f8"
+#    styles[Generic.Output] = "#355"
+#    styles[Generic.Prompt] = "bold #346ec9"
 
 
 dist = pkg_resources.Distribution(__file__)
@@ -94,7 +97,7 @@ ep = pkg_resources.EntryPoint.parse("hatchet = conf:HatchetStyle", dist=dist)
 dist._ep_map = {"pygments.styles": {"plugin1": ep}}
 pkg_resources.working_set.add(dist)
 
-pygments_style = "hatchet"
+#pygments_style = "hatchet"
 
 
 # -- Options for HTML output -------------------------------------------------
